+++ jvm/selfie-lib/src/commonMain/kotlin/com/diffplug/selfie/ArrayMap.kt
@@ -120,9 +120,9 @@ class ArrayMap<K : Comparable<K>, V : Any>(private val data: Array<Any>) : Map<K
   }
   /**
    * Returns an ArrayMap which has added or overwritten the given key/value. If the map already
-   * contained that mapping (equal keys and values) then it returns the same map.
+   * contained that mapping (equal keys and values) then it returns the identically same map.
    */
-  fun plusOrReplace(key: K, newValue: V): ArrayMap<K, V> {
+  fun plusOrNoOpOrReplace(key: K, newValue: V): ArrayMap<K, V> {
     val idxExisting = dataAsKeys.binarySearch(key)
     if (idxExisting >= 0) {
       val existingValue = data[idxExisting * 2 + 1] as V
@@ -137,6 +137,9 @@ class ArrayMap<K : Comparable<K>, V : Any>(private val data: Array<Any>) : Map<K
       return insert(idxInsert, key, newValue)
     }
   }
+
+  @Deprecated("Use plusOrNoOpOrReplace instead", ReplaceWith("plusOrNoOpOrReplace(key, newValue)"))
+  fun plusOrReplace(key: K, newValue: V) = plusOrNoOpOrReplace(key, newValue)
   private fun insert(idxInsert: Int, key: K, value: V): ArrayMap<K, V> {
     return when (data.size) {
       0 -> ArrayMap(arrayOf(key, value))
@@ -204,6 +207,22 @@ class ArrayMap<K : Comparable<K>, V : Any>(private val data: Array<Any>) : Map<K
   }
   override fun hashCode(): Int = entries.hashCode()
   override fun toString() = this.toMutableMap().toString()
+  fun minusOrNoOp(key: K, value: V): ArrayMap<K, V> {
+    val idxExisting = dataAsKeys.binarySearch(key)
+    if (idxExisting >= 0) {
+      val existingValue = data[idxExisting * 2 + 1] as V
+      if (value == existingValue) {
+        return minusSortedIndices(listOf(idxExisting))
+      }
+    }
+    return this
+  }
+  fun minusOrNoOp(key: K): ArrayMap<K, V> {
+    val idxExisting = dataAsKeys.binarySearch(key)
+    return if (idxExisting >= 0) {
+      minusSortedIndices(listOf(idxExisting))
+    } else this
+  }
 
   companion object {
     private val EMPTY = ArrayMap<String, Any>(arrayOf())