+++ jvm/selfie-lib/src/commonMain/kotlin/com/diffplug/selfie/guts/SnapshotSystem.kt
@@ -75,7 +75,7 @@ interface FS {
 }
 
 /** NOT FOR ENDUSERS. Implemented by Selfie to integrate with various test frameworks. */
-interface SnapshotStorage {
+interface SnapshotSystem {
   val fs: FS
   val mode: Mode
   val layout: SnapshotFileLayout
@@ -83,6 +83,14 @@ interface SnapshotStorage {
   fun sourceFileHasWritableComment(call: CallStack): Boolean
   /** Indicates that the following value should be written into test sourcecode. */
   fun writeInline(literalValue: LiteralValue<*>, call: CallStack)
+  /** Returns the DiskStorage for the test associated with this thread, else error. */
+  fun diskThreadLocal(): DiskStorage
+  /** Returns the DiskStorage for the test associated with this coroutine, else error. */
+  suspend fun diskCoroutine(): DiskStorage
+}
+
+/** Represents the disk storage for a specific test. */
+interface DiskStorage {
   /** Reads the given snapshot from disk. */
   fun readDisk(sub: String, call: CallStack): Snapshot?
   /** Writes the given snapshot to disk. */
@@ -94,7 +102,7 @@ interface SnapshotStorage {
   fun keep(subOrKeepAll: String?)
 }
 
-expect fun initStorage(): SnapshotStorage
+expect fun initSnapshotSystem(): SnapshotSystem
 
 interface SnapshotFileLayout {
   val rootFolder: TypedPath
@@ -102,4 +110,5 @@ interface SnapshotFileLayout {
   val allowMultipleEquivalentWritesToOneLocation: Boolean
   fun sourcePathForCall(call: CallLocation): TypedPath
   fun sourcePathForCallMaybe(call: CallLocation): TypedPath?
+  fun checkForSmuggledError()
 }