+++ jvm/undertest-junit5-kotest/src/test/kotlin/undertest/junit5/UT_KotestStringSpecTest.kt
@@ -1,0 +1,11 @@
+package undertest.junit5
+
+import com.diffplug.selfie.Selfie.expectSelfie
+import io.kotest.core.spec.style.StringSpec
+
+class UT_KotestStringSpecTest :
+    StringSpec({
+      "a" { expectSelfie("a").toMatchDisk() }
+      "b" { expectSelfie("b").toMatchDisk() }
+      "c" { expectSelfie("c").toMatchDisk() }
+    })